requirements.txt
langchain-core
langgraph
langchain-groq
python-dotenv
requests